+++ apps/web/src/pages/api/webhooks/stripe/[publickey].ts
@@ -1,0 +1,10 @@
+import { stripeHandler } from "@dotkomonline/api"
+
+export default stripeHandler
+
+// Required by stripe
+export const config = {
+  api: {
+    bodyParser: false,
+  },
+}
