+++ packages/api/index.ts
@@ -6,3 +6,4 @@ export type { Context } from "./src/context"
 export { transformer } from "./src/transformer"
 
 export { clerkHandler } from "./src/handlers/clerk-webhook"
+export { stripeHandler } from "./src/handlers/stripe-webhook"
