+++ packages/api/src/utils/request-utils.ts
@@ -1,0 +1,17 @@
+import { NextApiRequest } from "next"
+
+export const bufferRequest = (req: NextApiRequest) => {
+  return new Promise<Buffer>((resolve, reject) => {
+    const chunks: Buffer[] = []
+
+    req.on("data", (chunk: Buffer) => {
+      chunks.push(chunk)
+    })
+
+    req.on("end", () => {
+      resolve(Buffer.concat(chunks))
+    })
+
+    req.on("error", reject)
+  })
+}
